refactor(formatter): Move oxfmt utils (#21666)
Resolve TODO comment.

> // TODO: This logic should(can) move to this file, after LSP support is also moved here.
